vm - moved static and single tx vm hooks to scenarios, cleanup
