+++ chain/vm/src/vm_hooks.rs
@@ -1,9 +1,9 @@
 mod vh_dispatcher;
 mod vh_handler;
-mod vh_impl;
 mod vh_source;
+mod vh_tx_context;
 
 pub use vh_dispatcher::VMHooksDispatcher;
 pub use vh_handler::*;
-pub use vh_impl::*;
 pub use vh_source::VMHooksHandlerSource;
+pub use vh_tx_context::TxContextVMHooksHandler;
