--- chain/vm/src/vm_hooks/vh_impl.rs
@@ -1,9 +1,0 @@
-mod vh_debug_api;
-mod vh_single_tx_api;
-mod vh_static_api;
-mod vh_tx_context;
-
-pub use vh_debug_api::DebugApiVMHooksHandler;
-pub use vh_single_tx_api::{SingleTxApiData, SingleTxApiVMHooksHandler};
-pub use vh_static_api::StaticApiVMHooksHandler;
-pub use vh_tx_context::TxContextVMHooksHandler;
