+++ framework/scenario/src/api/impl_vh.rs
@@ -2,6 +2,8 @@ mod debug_api;
 mod debug_handle_vh;
 mod single_tx_api;
 mod static_api;
+mod vh_single_tx_api;
+mod vh_static_api;
 mod vm_hooks_api;
 mod vm_hooks_backend;
 
@@ -9,6 +11,8 @@ pub use debug_api::{DebugApi, DebugApiBackend};
 pub use debug_handle_vh::DebugHandle;
 pub use single_tx_api::SingleTxApi;
 pub use static_api::StaticApi;
+pub use vh_single_tx_api::{SingleTxApiData, SingleTxApiVMHooksHandler};
+pub use vh_static_api::StaticApiVMHooksHandler;
 pub(crate) use vm_hooks_api::i32_to_bool;
 pub use vm_hooks_api::VMHooksApi;
 pub use vm_hooks_backend::VMHooksApiBackend;
