+++ framework/scenario/src/api/impl_vh/single_tx_api.rs
@@ -1,16 +1,13 @@
 use std::sync::Mutex;
 
 use multiversx_chain_vm::{
-    executor::VMHooks,
-    types::VMAddress,
-    vm_hooks::{SingleTxApiData, SingleTxApiVMHooksHandler, VMHooksDispatcher},
-    world_mock::AccountData,
+    executor::VMHooks, types::VMAddress, vm_hooks::VMHooksDispatcher, world_mock::AccountData,
 };
 use multiversx_sc::api::RawHandle;
 
 use crate::debug_executor::StaticVarData;
 
-use super::{VMHooksApi, VMHooksApiBackend};
+use super::{SingleTxApiData, SingleTxApiVMHooksHandler, VMHooksApi, VMHooksApiBackend};
 
 thread_local! {
     static SINGLE_TX_API_VH_CELL: Mutex<SingleTxApiVMHooksHandler> = Mutex::default();
