+++ framework/scenario/src/api/impl_vh/vh_static_api.rs
@@ -1,9 +1,9 @@
 use std::sync::{Mutex, MutexGuard};
 
-use multiversx_chain_core::types::ReturnCode;
 use multiversx_chain_vm_executor::{MemLength, MemPtr};
 
-use crate::{
+use multiversx_chain_vm::{
+    chain_core::types::ReturnCode,
     tx_mock::{BackTransfers, TxFunctionName, TxInput, TxLog, TxManagedTypes, TxResult},
     types::{VMAddress, VMCodeMetadata},
     vm_hooks::{
@@ -15,7 +15,7 @@ use crate::{
     world_mock::{AccountData, BlockInfo},
 };
 
-use super::DebugApiVMHooksHandler;
+use crate::debug_executor::ContractDebugInstance;
 
 /// A simple wrapper around a managed type container Mutex.
 ///
@@ -31,13 +31,13 @@ impl StaticApiVMHooksHandler {
 impl VMHooksHandlerSource for StaticApiVMHooksHandler {
     unsafe fn memory_load(&self, offset: MemPtr, length: MemLength) -> &[u8] {
         // TODO: switch to the DebugSCInstance method
-        unsafe { DebugApiVMHooksHandler::main_memory_load(offset, length) }
+        unsafe { ContractDebugInstance::main_memory_load(offset, length) }
     }
 
     unsafe fn memory_store(&self, offset: MemPtr, data: &[u8]) {
         // TODO: switch to the DebugSCInstance method
         unsafe {
-            DebugApiVMHooksHandler::main_memory_store(offset, data);
+            ContractDebugInstance::main_memory_store(offset, data);
         }
     }
